updated app to newest command line interface
