+++ harvesting/app.py
@@ -1,14 +1,13 @@
 import click
 import logging
 import sys
-
 from util import db_connect
-from cache_software import cache_software
-from corporate import get_projects
-from github import sync_all as get_commits
 from releases import get_citations
+from github import sync_all as get_commits
 from zotero import get_mentions
 from oaipmh import list_records
+from corporate import get_projects
+from cache_software import cache_software
 
 
 class MaxLevel(object):
@@ -33,43 +32,69 @@ logger.setLevel(logging.INFO)
 logger.addHandler(stdout_handler)
 logger.addHandler(stderr_handler)
 
-@click.group()
+
+@click.group('cli')
 def cli():
+    """Command line interface to harvest data from various resources and combine the 
+    relevant information into one document."""
     pass
 
-harvest_choices = click.Choice(['commits', 'mentions', 'projects', 'citations', 'metadata', 'all'])
-@cli.command('harvest', help='Specify the type of data to be harvested.')
-@click.argument('type_of_data', type=harvest_choices, required=True)
-def harvest(type_of_data):
-    db = db_connect()
-    if type_of_data == 'commits':
-        get_commits()
-    elif type_of_data == 'mentions':
-        get_mentions()
-    elif type_of_data == 'projects':
-        get_projects()
-    elif type_of_data == 'citations':
-        get_citations(db)
-    elif type_of_data == 'metadata':
-        list_records()
-    elif type_of_data == 'all':
-        get_commits()
-        get_citations(db)
-        get_mentions()
-        get_projects()
-        list_records()
-    else:
-        raise Exception("I don't know how to harvest '" + harvest + "'.")
+
+@cli.group('harvest')
+def harvest_group():
+    """Harvest data from a variety of sources"""
+    pass
 
 
-@cli.command('resolve', help='Resolve foreign keys')
+@harvest_group.command('commits')
+def harvest_commits():
+    """Harvest commit data using GitHub's API"""
+    get_commits()
+
+
+@harvest_group.command('mentions')
+@click.option('--since-version', 'since_version', type=int)
+@click.option('--keys', 'keys', type=str)
+def harvest_mentions(since_version=None, keys=None):
+    """Harvest mentions from Zotero"""
+    get_mentions(since_version=since_version, keys=keys)
+
+
+@harvest_group.command('projects')
+def harvest_projects():
+    """Harvest project descriptions from https://esciencecenter.nl/projects"""
+    get_projects()
+
+
+@harvest_group.command('citations')
+def harvest_citations():
+    """Harvest citation metadata using Zenodo.org and GitHub.com APIs"""
+    get_citations(db)
+
+
+@harvest_group.command('metadata')
+def harvest_metadata():
+    """harvest metadata"""
+    list_records()
+
+
+@harvest_group.command('all')
+def harvest_all():
+    """Harvest commits, citations, mentions, projects"""
+    get_commits()
+    get_citations(db)
+    get_mentions(since_version=None)
+    get_projects()
+    list_records()
+
+
+@cli.command('resolve')
 def resolve():
-    db = db_connect()
+    """Combine information from different collections
+    into one document by resolving foreign keys"""
     cache_software()
 
-cli = click.CommandCollection(sources=[cli])
 
 if __name__ == '__main__':
+    db = db_connect()
     cli()
-
-
